fix(settings): prevent account deletion without deleting restaurants

+++ app/actions/settings/delete_account.ts
@@ -1,7 +1,8 @@
 import Restaurant from '#models/restaurant'
 import User from '#models/user'
 import db from '@adonisjs/lucid/services/db'
-import { TransactionClientContract } from '@adonisjs/lucid/types/database'
+import { Role } from '#enums/role'
+import ForbiddenException from '#exceptions/forbidden_exception'
 
 type Params = {
   user: User
@@ -12,15 +13,18 @@ export default class DeleteAccount {
     await db.transaction(async (trx) => {
       user.useTransaction(trx)
 
-      await this.#deleteDanglingRestaurants(user, trx)
+      const isAdminInRestaurant = await Restaurant.query({ client: trx })
+        .whereHas('users', (query) => {
+          query.where('users.id', user.id)
+          query.where('restaurant_users.role_id', Role.Admin)
+        })
+        .first()
+
+      if (isAdminInRestaurant) {
+        throw new ForbiddenException()
+      }
+
       await user.delete()
     })
-  }
-
-  static async #deleteDanglingRestaurants(user: User, trx: TransactionClientContract) {
-    return Restaurant.query({ client: trx })
-      .whereHas('users', (query) => query.where('users.id', user.id))
-      .whereDoesntHave('users', (query) => query.whereNot('users.id', user.id))
-      .delete()
   }
 }
